requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0

# Detection
ultralytics>=8.0.0

# ONNX
onnx>=1.14.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0  # Optional: GPU inference

# Visualization
matplotlib>=3.7.0
opencv-python>=4.8.0

# Demo
gradio>=4.0.0

# Data
roboflow>=1.1.0
PyYAML>=6.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
